Moved the service injection code from Main.cs file to a separate file.
